LOOP-2321 Move upgrades not on current release to 106
