+++ portlets/asset-entry-set-portlet/docroot/WEB-INF/src/com/liferay/asset/entry/set/hook/upgrade/v1_0_6/UpgradeAssetEntrySet.java
@@ -12,23 +12,17 @@
  * details.
  */
 
-package com.liferay.asset.entry.set.hook.upgrade.v1_0_5;
+package com.liferay.asset.entry.set.hook.upgrade.v1_0_6;
 
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
-import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
 
 /**
  * @author Matthew Kong
@@ -37,10 +31,10 @@ public class UpgradeAssetEntrySet extends UpgradeProcess {
 
 	@Override
 	protected void doUpgrade() throws Exception {
-		updateImageData();
+		updateLinkData();
 	}
 
-	protected void updateImageData() throws Exception {
+	protected void updateLinkData() throws Exception {
 		Connection con = null;
 		PreparedStatement ps = null;
 		ResultSet rs = null;
@@ -62,64 +56,16 @@ public class UpgradeAssetEntrySet extends UpgradeProcess {
 				JSONObject payloadJSONObject = JSONFactoryUtil.createJSONObject(
 					payload);
 
-				if (!Validator.equals(
-						payloadJSONObject.getString("type"), "image")) {
+				String linkData = payloadJSONObject.getString("linkData");
 
+				if (Validator.isNull(linkData)) {
 					continue;
 				}
 
-				JSONArray newImageDataJSONArray =
-					JSONFactoryUtil.createJSONArray();
+				JSONObject linkDataJSONObject =
+					JSONFactoryUtil.createJSONObject(linkData);
 
-				JSONArray imageDataJSONArray = payloadJSONObject.getJSONArray(
-					"imageData");
-
-				boolean modified = false;
-
-				for (int i = 0; i < imageDataJSONArray.length(); i++) {
-					JSONObject imageJSONObject =
-						imageDataJSONArray.getJSONObject(i);
-
-					String fullImageURL = imageJSONObject.getString(
-						"imageURL_full");
-
-					if (Validator.isNotNull(fullImageURL)) {
-						newImageDataJSONArray.put(imageJSONObject);
-
-						continue;
-					}
-
-					Map<String, String> newData = new HashMap<String, String>();
-
-					Iterator<String> iterator = imageJSONObject.keys();
-
-					while (iterator.hasNext()) {
-						String key = iterator.next();
-
-						if (!key.contains("raw")) {
-							continue;
-						}
-
-						String newKey = StringUtil.replace(key, "raw", "full");
-						String value = imageJSONObject.getString(key);
-
-						newData.put(newKey, value);
-					}
-
-					for (String key : newData.keySet()) {
-						imageJSONObject.put(key, newData.get(key));
-					}
-
-					newImageDataJSONArray.put(imageJSONObject);
-
-					modified = true;
-				}
-
-				if (!modified) {
-					continue;
-				}
-
-				payloadJSONObject.put("imageData", newImageDataJSONArray);
+				payloadJSONObject.put("linkData", linkDataJSONObject);
 
 				ps.setString(1, payloadJSONObject.toString());
