+++ portlets/asset-entry-set-portlet/docroot/WEB-INF/src/portal.properties
@@ -1,4 +1,4 @@
-release.info.build.number=105
+release.info.build.number=106
 release.info.previous.build.number=0
 
 upgrade.processes=\
@@ -7,6 +7,7 @@ upgrade.processes=\
     com.liferay.asset.entry.set.hook.upgrade.UpgradeProcess_1_0_2,\
     com.liferay.asset.entry.set.hook.upgrade.UpgradeProcess_1_0_3,\
     com.liferay.asset.entry.set.hook.upgrade.UpgradeProcess_1_0_4,\
-    com.liferay.asset.entry.set.hook.upgrade.UpgradeProcess_1_0_5
+    com.liferay.asset.entry.set.hook.upgrade.UpgradeProcess_1_0_5,\
+    com.liferay.asset.entry.set.hook.upgrade.UpgradeProcess_1_0_6
 
 application.startup.events=com.liferay.asset.entry.set.hook.events.StartupAction
